renamed file FolderQuestion to SmartFolderField
