+++ src/UI/DynamicFormBuilder.jsx
@@ -6,7 +6,7 @@ import React from "react"
 import { Button, ButtonGroup, Typography } from "@mui/material"
 import { capitalizeFirstLetter } from "../helperFunctions/OtherHelper"
 import { Field, Form, Formik } from "formik"
-import { SmartFolderField } from "./FormComponents/FieldTypes/FolderQuestion"
+import { SmartFolderField } from "./FormComponents/FieldTypes/SmartFolderField"
 
 const nestedFolders = [
   "models",
